+++ nautilus_trader/adapters/ccxt/data.pyx
@@ -23,7 +23,6 @@ from nautilus_trader.adapters.ccxt.providers import CCXTInstrumentProvider
 
 from nautilus_trader.common.clock cimport LiveClock
 from nautilus_trader.common.logging cimport Logger
-from nautilus_trader.core.constants cimport *  # str constants only
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.datetime cimport dt_to_unix_millis
 from nautilus_trader.core.datetime cimport millis_to_nanos