+++ nautilus_trader/model/commands.pyx
@@ -15,9 +15,10 @@
 
 from libc.stdint cimport uint64_t
 
+import json
+
 from nautilus_trader.core.uuid cimport UUID
-from nautilus_trader.model.identifiers cimport AccountId
-from nautilus_trader.model.identifiers cimport ClientId
+from nautilus_trader.model.events cimport OrderInitialized
 from nautilus_trader.model.identifiers cimport InstrumentId
 from nautilus_trader.model.identifiers cimport PositionId
 from nautilus_trader.model.identifiers cimport StrategyId
@@ -25,6 +26,7 @@ from nautilus_trader.model.identifiers cimport TraderId
 from nautilus_trader.model.objects cimport Price
 from nautilus_trader.model.orders.base cimport Order
 from nautilus_trader.model.orders.bracket cimport BracketOrder
+from nautilus_trader.model.orders.unpacker cimport OrderUnpacker
 
 
 cdef class TradingCommand(Command):
@@ -127,6 +129,55 @@ cdef class SubmitOrder(TradingCommand):
                 f"strategy_id={self.strategy_id.value}, "
                 f"command_id={self.id})")
 
+    @staticmethod
+    cdef SubmitOrder from_dict_c(dict values):
+        cdef str p = values["position_id"]
+        cdef PositionId position_id = PositionId(p) if p is not None else None
+        return SubmitOrder(
+            trader_id=TraderId(values["trader_id"]),
+            strategy_id=StrategyId(values["strategy_id"]),
+            position_id=position_id,
+            order=OrderUnpacker.unpack_c(json.loads(values["order"])),
+            command_id=UUID.from_str_c(values["command_id"]),
+            timestamp_ns=values["timestamp_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return a submit order command from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        SubmitOrder
+
+        """
+        return SubmitOrder.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "trader_id": self.trader_id.value,
+            "strategy_id": self.strategy_id.value,
+            "position_id": self.position_id.value if self.position_id is not None else None,
+            "order": json.dumps(self.order.init_event_c().to_dict()),
+            "command_id": self.id.value,
+            "timestamp_ns": self.timestamp_ns,
+        }
+
 
 cdef class SubmitBracketOrder(TradingCommand):
     """
@@ -181,6 +232,58 @@ cdef class SubmitBracketOrder(TradingCommand):
                 f"client_order_link_id={self.bracket_order.id.value}, "
                 f"command_id={self.id})")
 
+    @staticmethod
+    cdef SubmitBracketOrder from_dict_c(dict values):
+        cdef BracketOrder bracket_order = BracketOrder(
+            entry=OrderUnpacker.unpack_c(json.loads(values["entry"])),
+            stop_loss=OrderUnpacker.unpack_c(json.loads(values["stop_loss"])),
+            take_profit=OrderUnpacker.unpack_c(json.loads(values["take_profit"])),
+        )
+        return SubmitBracketOrder(
+            trader_id=TraderId(values["trader_id"]),
+            strategy_id=StrategyId(values["strategy_id"]),
+            bracket_order=bracket_order,
+            command_id=UUID.from_str_c(values["command_id"]),
+            timestamp_ns=values["timestamp_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return a submit bracket order command from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        SubmitBracketOrder
+
+        """
+        return SubmitOrder.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "trader_id": self.trader_id.value,
+            "strategy_id": self.strategy_id.value,
+            "entry": json.dumps(self.bracket_order.entry.init_event_c().to_dict()),
+            "stop_loss": json.dumps(self.bracket_order.stop_loss.init_event_c().to_dict()),
+            "take_profit": json.dumps(self.bracket_order.take_profit.init_event_c().to_dict()),
+            "command_id": self.id.value,
+            "timestamp_ns": self.timestamp_ns,
+        }
+
 
 cdef class UpdateOrder(TradingCommand):
     """
@@ -199,8 +302,9 @@ cdef class UpdateOrder(TradingCommand):
         InstrumentId instrument_id not None,
         ClientOrderId client_order_id not None,
         VenueOrderId venue_order_id not None,
-        Quantity quantity not None,
-        Price price not None,
+        Quantity quantity,  # Can be None
+        Price price,  # Can be None
+        Price trigger,  # Can be None
         UUID command_id not None,
         uint64_t timestamp_ns,
     ):
@@ -219,10 +323,12 @@ cdef class UpdateOrder(TradingCommand):
             The client order identifier to update.
         venue_order_id : VenueOrderId
             The venue order identifier to update.
-        quantity : Quantity
-            The quantity for the order (update optional).
-        price : Price
-            The price for the order (update optional).
+        quantity : Quantity, optional
+            The quantity for the order update.
+        price : Price, optional
+            The price for the order update.
+        trigger : Price, optional
+            The trigger price for the order update.
         command_id : UUID
             The command identifier.
         timestamp_ns : uint64
@@ -241,6 +347,7 @@ cdef class UpdateOrder(TradingCommand):
         self.venue_order_id = venue_order_id
         self.quantity = quantity
         self.price = price
+        self.trigger = trigger
 
     def __repr__(self) -> str:
         return (f"{type(self).__name__}("
@@ -251,7 +358,66 @@ cdef class UpdateOrder(TradingCommand):
                 f"venue_order_id={self.venue_order_id.value}, "
                 f"quantity={self.quantity.to_str()}, "
                 f"price={self.price}, "
+                f"trigger={self.trigger}, "
                 f"command_id={self.id})")
+
+    @staticmethod
+    cdef UpdateOrder from_dict_c(dict values):
+        cdef str q = values["quantity"]
+        cdef str p = values["price"]
+        cdef str t = values["trigger"]
+        return UpdateOrder(
+            trader_id=TraderId(values["trader_id"]),
+            strategy_id=StrategyId(values["strategy_id"]),
+            instrument_id=InstrumentId.from_str_c(values["instrument_id"]),
+            client_order_id=ClientOrderId(values["client_order_id"]),
+            venue_order_id=VenueOrderId(values["venue_order_id"]),
+            quantity=Quantity.from_str_c(q) if q is not None else None,
+            price=Price.from_str_c(p) if p is not None else None,
+            trigger=Price.from_str_c(t) if t is not None else None,
+            command_id=UUID.from_str_c(values["command_id"]),
+            timestamp_ns=values["timestamp_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return an update order command from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        UpdateOrder
+
+        """
+        return SubmitOrder.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "trader_id": self.trader_id.value,
+            "strategy_id": self.strategy_id.value,
+            "instrument_id": self.instrument_id.value,
+            "client_order_id": self.client_order_id.value,
+            "venue_order_id": self.venue_order_id.value,
+            "quantity": str(self.quantity) if self.quantity is not None else None,
+            "price": str(self.price) if self.price is not None else None,
+            "trigger": str(self.trigger) if self.trigger is not None else None,
+            "command_id": self.id.value,
+            "timestamp_ns": self.timestamp_ns,
+        }
 
 
 cdef class CancelOrder(TradingCommand):
@@ -314,3 +480,52 @@ cdef class CancelOrder(TradingCommand):
                 f"client_order_id={self.client_order_id.value}, "
                 f"venue_order_id={self.venue_order_id.value}, "
                 f"command_id={self.id})")
+
+    @staticmethod
+    cdef CancelOrder from_dict_c(dict values):
+        return CancelOrder(
+            trader_id=TraderId(values["trader_id"]),
+            strategy_id=StrategyId(values["strategy_id"]),
+            instrument_id=InstrumentId.from_str_c(values["instrument_id"]),
+            client_order_id=ClientOrderId(values["client_order_id"]),
+            venue_order_id=VenueOrderId(values["venue_order_id"]),
+            command_id=UUID.from_str_c(values["command_id"]),
+            timestamp_ns=values["timestamp_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return a cancel order command from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        CancelOrder
+
+        """
+        return CancelOrder.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "trader_id": self.trader_id.value,
+            "strategy_id": self.strategy_id.value,
+            "instrument_id": self.instrument_id.value,
+            "client_order_id": self.client_order_id.value,
+            "venue_order_id": self.venue_order_id.value,
+            "command_id": self.id.value,
+            "timestamp_ns": self.timestamp_ns,
+        }